+++ SwiftExample/ViewController.swift
@@ -5,12 +5,151 @@
 
 import UIKit
 import QuartzCore
+import ImageIO
+
+class ArtworkRequest: NSObject, NSURLConnectionDataDelegate {
+    let maximumResponseSize = 1024 * 1024
+    var data = NSMutableData()
+    var responseAccepted = false
+    var requestCompleted = false
+    var connection: NSURLConnection?
+    let completion: (NSData?) -> Void
+
+    init?(URL: NSURL, completion: (NSData?) -> Void) {
+        self.completion = completion
+        super.init()
+
+        let request = NSURLRequest(
+            URL: URL,
+            cachePolicy: NSURLRequestCachePolicy.ReloadIgnoringLocalCacheData,
+            timeoutInterval: 15
+        )
+        guard let connection = NSURLConnection(request: request, delegate: self, startImmediately: false) else {
+            return nil
+        }
+
+        self.connection = connection
+    }
+
+    func start() {
+        connection?.start()
+    }
+
+    func cancel() {
+        connection?.cancel()
+        completeWithData(nil)
+    }
+
+    class func isTrustedURL(URL: NSURL) -> Bool {
+        guard URL.user == nil && URL.password == nil && URL.port == nil && URL.fragment == nil else {
+            return false
+        }
+
+        if let scheme = URL.scheme?.lowercaseString,
+            host = URL.host?.lowercaseString {
+                return scheme == "https" && (host == "mzstatic.com" || host.hasSuffix(".mzstatic.com"))
+        }
+
+        return false
+    }
+
+    func isAcceptableResponse(response: NSURLResponse) -> Bool {
+        guard let httpResponse = response as? NSHTTPURLResponse where
+            httpResponse.statusCode >= 200 && httpResponse.statusCode < 300 else {
+                return false
+        }
+
+        guard let responseURL = response.URL where ArtworkRequest.isTrustedURL(responseURL) else {
+            return false
+        }
+
+        let contentLength = response.expectedContentLength
+        if contentLength > Int64(maximumResponseSize) {
+            return false
+        }
+
+        guard let mimeType = response.MIMEType?.lowercaseString else {
+            return false
+        }
+
+        return mimeType == "image/jpeg" || mimeType == "image/png"
+    }
+
+    func canAppendArtworkData(chunk: NSData) -> Bool {
+        return responseAccepted && chunk.length <= maximumResponseSize - data.length
+    }
+
+    func completeWithData(result: NSData?) {
+        if requestCompleted {
+            return
+        }
+
+        requestCompleted = true
+        connection = nil
+        data = NSMutableData()
+        completion(result)
+    }
+
+    func connection(connection: NSURLConnection, didFailWithError error: NSError) {
+        if self.connection !== connection {
+            return
+        }
+
+        completeWithData(nil)
+    }
+
+    func connection(connection: NSURLConnection, didReceiveResponse response: NSURLResponse) {
+        if self.connection !== connection {
+            return
+        }
+
+        data = NSMutableData()
+        responseAccepted = isAcceptableResponse(response)
+        if !responseAccepted {
+            connection.cancel()
+            completeWithData(nil)
+        }
+    }
+
+    func connection(connection: NSURLConnection, didReceiveData chunk: NSData) {
+        if self.connection !== connection {
+            return
+        }
+
+        if !canAppendArtworkData(chunk) {
+            connection.cancel()
+            completeWithData(nil)
+            return
+        }
+
+        data.appendData(chunk)
+    }
+
+    func connectionDidFinishLoading(connection: NSURLConnection) {
+        if self.connection !== connection {
+            return
+        }
+
+        if !responseAccepted {
+            completeWithData(nil)
+            return
+        }
+
+        completeWithData(NSData(data: data))
+    }
+}
 
 class SearchResultsViewController: UIViewController, UITableViewDataSource, UITableViewDelegate, APIControllerProtocol {
     
     var api: APIController = APIController()
     @IBOutlet var appsTableView : UITableView?
     var tableData: NSArray = NSArray()
+    let maximumArtworkDimension = 8192
+    let maximumArtworkPixelCount = 16 * 1024 * 1024
+    let maximumArtworkURLLength = 2048
+    let maximumResultCount = 200
+    var artworkGeneration = 0
+    var artworkRequests = [NSIndexPath: ArtworkRequest]()
     
     override func viewDidLoad() {
         super.viewDidLoad()
@@ -21,7 +160,16 @@ class SearchResultsViewController: UIViewController, UITableViewDataSource, UITa
 
     override func viewWillDisappear(animated: Bool) {
         super.viewWillDisappear(animated)
+        api.cancel()
+        cancelArtworkRequests()
         UIApplication.sharedApplication().networkActivityIndicatorVisible = false
+    }
+
+    func cancelArtworkRequests() {
+        for request in artworkRequests.values {
+            request.cancel()
+        }
+        artworkRequests.removeAll()
     }
     
     func tableView(tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
@@ -37,8 +185,7 @@ class SearchResultsViewController: UIViewController, UITableViewDataSource, UITa
                 cell.textLabel?.text = rowData["trackName"] as? String
                 cell.imageView?.image = nil
 
-                if let urlString = rowData["artworkUrl60"] as? String,
-                    imgURL = safeArtworkURLFromString(urlString) {
+                if let imgURL = artworkURLForRow(indexPath) {
                         loadArtworkFromURL(imgURL, forCell: cell, tableView: tableView, indexPath: indexPath)
                 }
             }
@@ -61,31 +208,91 @@ class SearchResultsViewController: UIViewController, UITableViewDataSource, UITa
     }
 
     func safeArtworkURLFromString(urlString: String) -> NSURL? {
-        if let url = NSURL(string: urlString),
-            scheme = url.scheme?.lowercaseString,
-            host = url.host?.lowercaseString {
-                if scheme == "https" && (host == "mzstatic.com" || host.hasSuffix(".mzstatic.com")) {
-                    return url
-                }
+        guard urlString.utf8.count <= maximumArtworkURLLength else {
+            return nil
+        }
+
+        if let url = NSURL(string: urlString) {
+            if ArtworkRequest.isTrustedURL(url) {
+                return url
+            }
         }
 
         return nil
     }
 
+    func artworkURLForRow(indexPath: NSIndexPath) -> NSURL? {
+        guard indexPath.row < tableData.count,
+            let rowData = tableData[indexPath.row] as? NSDictionary,
+            urlString = rowData["artworkUrl60"] as? String else {
+                return nil
+        }
+
+        return safeArtworkURLFromString(urlString)
+    }
+
+    func canDisplayArtworkDimensions(width: Int, height: Int) -> Bool {
+        guard width > 0 && height > 0 &&
+            width <= maximumArtworkDimension && height <= maximumArtworkDimension else {
+            return false
+        }
+
+        return width <= maximumArtworkPixelCount / height
+    }
+
+    func isAcceptableArtworkImage(image: UIImage) -> Bool {
+        guard let cgImage = image.CGImage else {
+            return false
+        }
+
+        return canDisplayArtworkDimensions(
+            CGImageGetWidth(cgImage),
+            height: CGImageGetHeight(cgImage)
+        )
+    }
+
+    func isAcceptableArtworkData(data: NSData) -> Bool {
+        guard let source = CGImageSourceCreateWithData(data, nil),
+            properties = CGImageSourceCopyPropertiesAtIndex(source, 0, nil) as NSDictionary?,
+            width = properties[kCGImagePropertyPixelWidth] as? NSNumber,
+            height = properties[kCGImagePropertyPixelHeight] as? NSNumber else {
+                return false
+        }
+
+        return canDisplayArtworkDimensions(width.integerValue, height: height.integerValue)
+    }
+
     func loadArtworkFromURL(imgURL: NSURL, forCell cell: UITableViewCell, tableView: UITableView, indexPath: NSIndexPath) {
-        dispatch_async(dispatch_get_global_queue(DISPATCH_QUEUE_PRIORITY_DEFAULT, 0)) {
-            guard let imgData = NSData(contentsOfURL: imgURL),
-                image = UIImage(data: imgData) else {
+        artworkRequests[indexPath]?.cancel()
+        let generation = artworkGeneration
+        if let request = ArtworkRequest(URL: imgURL, completion: { [weak self, weak cell, weak tableView] imgData in
+            guard let controller = self,
+                targetCell = cell,
+                targetTableView = tableView,
+                data = imgData else {
                     return
             }
 
-            dispatch_async(dispatch_get_main_queue()) {
-                if let visibleIndexPath = tableView.indexPathForCell(cell)
-                    where visibleIndexPath.section == indexPath.section && visibleIndexPath.row == indexPath.row {
-                        cell.imageView?.image = image
-                        cell.setNeedsLayout()
+            dispatch_async(dispatch_get_global_queue(DISPATCH_QUEUE_PRIORITY_DEFAULT, 0)) {
+                guard controller.isAcceptableArtworkData(data),
+                    image = UIImage(data: data) where controller.isAcceptableArtworkImage(image) else {
+                    return
+                }
+
+                dispatch_async(dispatch_get_main_queue()) {
+                    if controller.artworkGeneration == generation,
+                        let visibleIndexPath = targetTableView.indexPathForCell(targetCell)
+                        where visibleIndexPath.section == indexPath.section && visibleIndexPath.row == indexPath.row,
+                        currentArtworkURL = controller.artworkURLForRow(indexPath)
+                        where currentArtworkURL.isEqual(imgURL) {
+                            targetCell.imageView?.image = image
+                            targetCell.setNeedsLayout()
+                    }
                 }
             }
+        }) {
+            artworkRequests[indexPath] = request
+            request.start()
         }
     }
     
@@ -98,8 +305,11 @@ class SearchResultsViewController: UIViewController, UITableViewDataSource, UITa
         }
 
         // Store the results in our table data array
+        artworkGeneration += 1
+        cancelArtworkRequests()
         if let resultsArray = results["results"] as? NSArray {
-            self.tableData = resultsArray
+            let resultCount = min(resultsArray.count, maximumResultCount)
+            self.tableData = resultsArray.subarrayWithRange(NSMakeRange(0, resultCount))
         } else {
             self.tableData = NSArray()
         }